Refactor environment parsing to use common.ParseEnvVariable and move tests
- Replace internal/runner/environment.ParseEnvVariable usages with common.ParseEnvVariable
  across docs and runner packages
- Rename and relocate parser tests to internal/common/env_test.go for shared usage
- Consolidate filesystem contains test into internal/common/filesystem_test.go and remove
  duplicate file
- Add package comment and nolint directive to internal/common/errors.go
- Remove thin parser wrapper; call common.ParseEnvVariable directly from callers
