+++ internal/common/env_test.go
@@ -1,4 +1,4 @@
-package environment
+package common
 
 import (
 	"testing"
