+++ internal/common/errors.go
@@ -1,4 +1,6 @@
 // Package common provides shared utilities and error definitions used across multiple packages.
+//
+//nolint:revive // common is an appropriate name for shared utilities package
 package common
 
 import "errors"
